shiny_app/requirements.txt
# Paquete principal para crear la Shiny App
shiny
python-dotenv
# Paquetes para el chatbot
chatlas>=0.4.0
google-genai
shiny
tweepy
## Paquetes para poder realizar el web scraping generalista
scrapy 
psycopg2-binary  
requests         
beautifulsoup4
selenium
## Para la API de X antes Twitter 
tweepy
## Para la API de Google
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
## Para el análisis de sentimientos
spacy
transformers
vaderSentiment
SpacyTextBlob
nltk
## Para las gráficas
matplotlib
seaborn       
plotly          
tqdm